Toggleable xbox button layout for nintendo controllers, axis flip (#59)
Also controller class uses int now because truncation was fucking with the axis flip

and the example code has another parameter because it makes my life easier.

+++ Chroma/Input/Controller.cs
@@ -55,7 +55,7 @@ namespace Chroma.Input
             return driver.CanIgnoreAxisMotion(axis, axisValue);
         }
 
-        public static short GetRawAxisValue(int playerIndex, ControllerAxis axis)
+        public static int GetRawAxisValue(int playerIndex, ControllerAxis axis)
         {
             var driver = Get(playerIndex);
 
@@ -75,7 +75,7 @@ namespace Chroma.Input
             return driver.GetRawAxisValueNormalized(axis);
         }
 
-        public static short GetAxisValue(int playerIndex, ControllerAxis axis)
+        public static int GetAxisValue(int playerIndex, ControllerAxis axis)
         {
             var driver = Get(playerIndex);
